Multiple Item Types Test
Different Item Types can be added to the inventory and display.

Known Bug:
Inventory will create a node child for each addition to the inventory when it is supposed to increment an existing node. This behavior is not present in the UI however...

Known Design Issue:
Inventory item defaults need to be re-evaluated -> very buggy when new items are introduced, the carrot item resource was baked into too many nodes.